requirements.txt
flask
flask-cors
toml
colorama
requests
fastlogging
Jinja2
keybert
sqlmodel